Return an empty packet when command not supported
The GDB specification states that all unsupported commands should be
responded to with an empty packet.  This patch modifies the functions
handleMsg and handleQuery to return empty packets if an unknown
or unsupported command is received.  This fixes the problem where
the "C/C++ GDB Hardware Debugging" eclipse plug-in hangs when
attempting to debug.