Close interface to prevent hanging on windows
In board.py the function uninit only closes the interface if an
exception occurs in either target.resume() or transport.uninit().
This patch moves the finally block running interface.close to
the top try block so the interface always gets closed.
This fixes the hanging commonly seen on windows.

This problem was introduced in the commit
f55ecc9bf9d9cb3fd380ab29774c1618a4f18eda - "Fix a different python
hang (on Windows) when an error occurred during the flash.init()."