Allow target to be explicitly specified
This patch adds an extra parameter, target_override, to
getAllConnectedBoards and chooseBoard in mbed_board.py and to the gdb
server.  This overrides the default target allowing the same mbed
firmware to be used to debug more than just the built in chip.  This
is particularly useful on platforms which have a jtag connector for
external debugging such as the Freescale FRDM boards.  This patch also
fixes spelling errors in the gdb server.