Prevent flash security check from locking target
The function checkSecurityBits returns 0 when the flash settings would
lock the target processor.  This in turn causes programPage to return
without programming the current page.  The problem with this is
on Kinetis processors, if the security byte is erased it will lock the
part.

This patch renames checkSecurityBits to overrideSecurityBits.  Data
to be flashed is passed into this function which modifies the security
bits to a safe value, and then returns the new data.