Add resume parameter to Board.uninit
When uninitializing a board, it is not always desirable to let the
processor run.  This patch adds the parameter 'resume' to Board.uninit.
This allows the calling script to uninitialize the board without
letting it run.  This is useful if you want to keep the board halted
when running a series of scripts.