Modify resetStopOnReset to use vector catch
Modify the function resetStopOnReset to use the vector catch instead
of a breakpoint.   This allows flashing to succeeded on blank parts
with an invalid reset address in the vector table.  This also frees up
an extra breakpoint during the flash sequence.

This commit also remove the redundant reset in the
function setTargetState.