Fix SWD fault when connecting to locked Kinetis
When initializing a locked Kinetis processor CortexM.init must not
access system registers until the device is unlocked.  This patch
modifies the call to CortexM.init in the Kineits init so it does not
try and initialize data breakpoints.  This fixes a SWD fault when
running blank_test.py.

This problem was introduced in the commit
907940d7d739f8f90aa7c20356cedc520eddddf4 -
"Add support for data breakpoints"