Fix incorrect GDB step and resume behavior
When using GDB, a typical way of moving the execution address is to set
a breakpoint with the command 'break [line]' and then jump to that
address with the command 'jump [line]'.  This does not work because
the function maybeSkipBreakpoint will cause GDB to skip the breakpoint.
This patch fixes that problem by removing the function
maybeSkipBreakpoint.

GDB single stepping still works as expected and does not get stuck
when trying to step over breakpoints.  This is because the GDB
client automatically removes the breakpoint when stepping over it
and restores it after.

When using Eclipse as a GDB front end this fixes the problem where
where "Move to line" causes the device to begin executing code,
rather than just moving the PC.