Lower CPU usage
When I had a device halted under pyOCD, the CPU usage on my development
machine would sky rocket.  The effect of this was very noticeable when
running off of batteries.  This was due to pyOCD's usage of
non-blocking socket reads in a tight loop while it tries to read
packets sent from GDB.  I added a sleep() call to the run() method as
already used in the resume() method.  I set both of these sleeps to be
100 milliseconds. For me, I didn't find this decreased the perceived
responsiveness of pyOCD when I was using GDB and it greatly reduces the
CPU usage.