Disable stepping into interrupts by default
If an interrupt is pending when GDB issues a step command, the
processor will step into the interrupt handler.  This behavior
is typically not what the user intended.  This patch makes
the single step command disable interrupts by default when
stepping to correct this behavior.  This can be turned off with
the GDB switch -s.

This also fixes the behavior where letting the processor run with
an interrupt pending after hitting a breakpoint results in the
target breaking at the same breakpoint, having only executed the
interrupt handler and no new code on the current thread.