Better handling for getTargetXML
Previously I was setting the targetXML field from the CortexM::init()
method once it knew if there was a FPU or not.  This broke for some
targets which don't perform the FPU check in the CortexM::init() method
but perform it slightly later themselves.  Now the getTargetXML()
method checks the self.has_fpu field itself and returns the appropriate
XML string.