Replace hardfault breakpoint with vector catch
Remove the hardfault breakpoint code and use the vector catch
instead.  This frees up a breakpoint, prevents interference when
there are no breakpoints free and removes the need to read the
hardfault vector (which causes problems on the nrf51 and other
targets).

This patch also adds the ability to halt on system resets.  This
functionality is turned off by default.