Add SWD frequency option to GDB server
Add a GDB option for changing the SWD clock frequency.  The default
clock frequency is 1MHz.  Increasing the frequency further than the
default can lead to a 20 - 40% performance increase, but is limited by
HID communication and debugger firmware.  Decreasing the frequency
can allow GDB to debug to devices running on a slower clock
configuration.