Enable eraseAll functionality for LPC4330
Previously I worked around the fact that the flash algorithm code that
I puled from the LPC4330 CMSIS-DAP implementation would fail.  This
commit removes that work around and instead modifies the code in the
flash algorithm to set the S_FORCE_ERASE option flag instead of
S_VERIFY_ERASE.  This change allows for a successful eraseAll() to be
conducted.

I also modified the programPage() flash algorithm implementation by
having it use the S_CALLER_ERASE option flag since the pyOCD code is
always expected to erase a page before it programs it.  Previously the
programPage() implementation was expecting that it would have to erase
the page first.