Fix FPU register write
This is a fix for the following issue reported on github:
https://github.com/mbedmicro/pyOCD/issues/95

I had previously tested that the 'G' packet command worked correctly to
update the contents of the FPU registers.  I had originally removed
support for the 'p' and 'P' packet commands but then added the 'P'
command back in and not tested it properly with FPU registers and
just concentrated on the core integer registers.

This fix takes into account that GDB sends 64-bit double precision
d* registers to update with the 'P' packet whereas I had previously
implemented it incorrectly to expect 32-bit single precision s*
register updates.  Since the target XML describes the FPU registers as
16 double precision register values, GDB will always perform such
64-bit updates.