Improve dapTransfer speed by queuing packets
Send multiple transfer block commands before reading any responses.
This allows multiple packets to be buffered up, which improves
 transfer speeds.  On the K22F with firmware 0203 and a clock of 10MHz
this patch brings the data transfer rate from ~25,000 B/s (before) to
~45,000 B/s (after).

To support this the interface functions 'setPacketCount' and
'getPacketCount' were created.  These function control the number of
packets that can be written to the interface before returning data
needs to be read.