Improve pyusb speed by creating reader thread
This patch creates a dedicated RX thread for the pyusb_backend to
speed up transfers.

This patch also releases the device handle when the given board is
closed.  This allows the same script to close and then re-open the
same device multiple times, as the blank_test.py script does.