Fix monitor commands with OpenOCD eclipse plugin
The latest version of the OpenOCD plug in at the time of this patch,
2.6.1-201502281154, sends the unsupported 'reg' monitor command.
In return pyOCD sends a blank response indicating that this command is
unsupported.  This causes the OpenOCD plug in to immediately exit the
debugging session.

This patch changes the response from an empty string to a hex string
explaining that the command is unsupported.  This allows pyOCD to work
with the OpenOCD eclipse plug in.  This new behavior matches that of
OpenOCD, which also returns an error message when an unsupported
monitor command is sent.  Returning an error string is also the
recommended response by the document "Debugging with gdb
(Tenth Edition, for gdb version 7.6.50.20131005-cvs)".

Unrelated to this, this patch also fixes the formatting on the monitor
'help' command by only printing the command description at index 0
rather than the entire list.