Fix LPC sector erase
Fix LPC sector erase for the lpc11u24, lpc1768 and lpc800.  This patch
adds the key pc_erase_sector to the flash programming algorithm of
these targets.  The lpc800 also needed to have its flash address moved
in basic_test.py since this was outside rom and caused the sector erase
test to fail.  On the lpc11u24 and lpc1768 the sector erase size did
not match the sector program size, so as a temporary fix programPage
has been overridden.