Remap flash to user code after reset on LPC
On the LPC11u24, LPC800 and LPC1768 the vector table is mapped to the
boot loader flash on reset rather than user flash.  This patch remaps
the flash back to user code and sets the SP and PC when
resetStopOnReset is called.

This patch also updates the vector table checksum in the LPC800.