Compute CRCs to analyze flash faster
When deciding between chip erase programming and sector erase
programming an estimate for the number of pages with the same
data must be known.  Previously this was done by reading
the first 32 bytes of each page and comparing that to the new
data.  This gives a good approximation, but does not catch all
changes.

This patch adds a binary which computes the CRC of each sector that
is going to be programmed.  By calling flash.computeCrcs() a CRC
for each page is computed.  This both decreases analysis time and
greatly increases estimate accuracy.  This also prevents
unnecessarily checking pages that have differing data, since
this can be reliably determined from the CRC.