Use AHB IDR to determine auto increment size
Prior to this patch the auto increment page size needed to be set for
each target.  This patch removes that requirement and instead sets the
value based on the what the AHB supports.  This reduces maintenance
and makes mistakes less likely.

This patch also boosts speed on targets that had their auto increment
page size set to the wrong value.  In particular the K22F and K64F see
a >10% speed increase when running speed_test.py.