Add generic Cortex-M target
Add support for a generic Cortex-M target.  When specified as the
target, any Cortex-M processor can be debugged with gdb_server.py.
Flash cannot be programmed when using this target and gdb does
not report a memory map.  Programs can still be loaded into ram
and debugged from there.

This can be useful for debugging processors which pyOCD does
not support fully, but can be programmed with another tool
(mass storage, bootloader, ...).  This is also useful for debugging
flash algorithms from ram when adding support for new boards.