Prevent mon commands from leaving target running
The GDB server assumes that after each command the target will be
halted.  If this is not the case, following commands could fail.
With this patch, if monitor commands leaves the target running, pyOCD
will print an error message and halt the target itself. This patch also
prints out an error message if single stepping fails because the target
already running.  This is to make it more obvious  why the target is
not stepping.

In eclipse with the GNU ARM Eclipse plugin this problem would
occur if the reset command's type was not 'halt' and the reset
button was pressed multiple times in a row.  When in this state, trying
to resume or single step would do nothing, and appear to halt
at the same address.  Reading core register would also show the wrong
value.