Map boards to board info rather than to target
Map each board ID to a board info object rather than only to a target.
This make it possible to have a different test binary for each board.
Dedicated test binaries are important on boards that use the same
processor, but have buttons, lights or other hardware on different
pins.