Rebuild nrf51 flash algo and add erase sector
This patch rebuilds the nrf51 flash algorithm and adds the function
pc_erase_sector.  The flash algorithm was rebuild from the
Nordic Semiconductor CMSIS-DAP repository at the commit
1fdb49fb723846872f34c7dc2c87c44cb590ad65 -
"Added new functionality for detecting which nordic board is used
and providing the user with the correct mbed.htm file."

This patch also changes 'begin_data' from 0x20000200 to 0x20002000 to
make more room for page data and for the stack.  The original value
was not causing problems and this just gives extra space.