Fix programming when reset handler is non-thumb
When the reset handler is set to a non-thumb address (even address)
then flash programming fails.  This is because the processor is set to
ARM mode, which is not implemented for Cortex processors.  On the
first instruction executed in the flash algorithm the device will crash.

To fix this problem this patch explicitly sets the thumb bit when
configuring the target state for programming.

This patch also adds a test case to flash_test.py catch this problem
in the future if there are regressions.