Add support for the nRF51-DK
Add support for the nRF51-DK and change references from nRF51822 to
just nRF51.  The nRF51822 and nRF51422 have a common reference manual
and are similar enough to use the same flash algorithm.

The board test binary l1_nrf51-dk.bin was built using the uVision
example project "blinkey_blank_pca10028".

This patch also modifies cortex_test.py to allow the nrf51 to pass the
invalid memory test even when an exception is not thrown.  This is
because on the nrf51, invalid memory accesses by the debugger
return 0 rather than causing a transfer error.