Add option to debug flash algorithm
Add the function setFlashAlgoDebug to control if the flash algorithm
should be debugged.  When enabled the function callFunction turns on
the reset and vector catches to trap errors, and checks the following:
1. The PC points to breakpoint address
2. The stack pointer is restored at the end of the call
3. The frame pointer was not modified

To accomplish the above checking, this patch also adds the cortex_m
functions getVectorCatchFalse, getVectorCatchReset and changes
how the function callFunction handles initialization.

fixup