Update handling of GDB breakpoints
Change the default behavior of the GDB software breakpoint command to
return an empty response.  This allows the GDB client to implement
software breakpoints by modifying ram addresses itself.  To force the
GDB server to implement software breakpoints using hardware breakpoints
the command line option '-k' can be used.

This patch also improves the handling for breakpoints set outside
the supported breakpoint region.  Now, instead of masking off
the higher bits and setting the breakpoint at the wrong address,
an error message is printed and the no breakpoint is set.