Add fast verify programming option
Add the option --fast_program to gdb_server.py and flash_tool.py.
When specified the flash programming code will use the CRC32 only
to determine if a page has identical data and does not need to be
programmed.  This is different from the normal behavior, where each
page with a matching CRC32 is read to confirm that the contents are
the same.  By bypassing this read, loading an image that is already on
a device occurs very fast, typically under 1s.

This patch also adds a test case to flash_test.py to verify that fast
verify works as expected.