install: declare dependencies for platforms in setup.py
This change declares the dependencies for the project in setup.py
so that users do not need to remember to install each of them.  Tools
like pip will parse the `install_requires` list and download and
install the packages for users.  I verified that each of the dependencies
does appear to be present on pypi already.