build: update/simplify instructions for building pyinstaller packages
Specific instructions are not provided for Windows, but there are no
major leaps required, so I think it makes sense to keep things simple.
The latest PyPI release of pyisntaller is now incredibly old and
the resulting binaries do not work with the released version, so
use of the upstream pyinstaller is required.