gdb: install a 'pyocd' executable on installation that runs gdb server
This change moves the main content from the gdb_server.py tool into
the main codebase and adds a console entry_point to the python package
definition.  The entry_points specification is well-understood by
tools like pip and is also familiar to package maintiners for various
distributions.

Some cleanup (but not complete cleanup) was performed on the code
that was moved, but everything should be functionally equivalent to
how it was previously.