docs: update readme with additional installation information
This includes basic instructions for installing from source
as well as information for setting up a development virtualenv.

Right now the dev-requirements.txt is very simple, but its
existence opens new possibilities with regards to adding additional
development dependencies (for instance, nose and mock could be added
in order to help with unit testing in the future).