tests: add initial unittests and related infrastructure
These changes add tests for some functions for the conversion
utilities.  Since some of the names for these functions were
not very good, those were also updated to be more meaningful.
Finally,  the implementation for intToHex8 was modified to be
much simpler (after writing a test that covered the code with
the existing implementation).

Signed-off-by: Paul Osborne <osbpau@gmail.com>