Add script for GDB testing
This patch adds the script gdb_test.py to allow for automated gdb
testing.  This script starts both a gdb host and a gdb client
which communicate and simulate a gdb session.  In this
session a program is loaded into ram to allow for further testing.

Some things tested in this script:
-Continue & halt work
-monitor commands can be sent
-memory can be written/read
-reading invalid memory does not cause a crash
-setting too many breakpoints/watchpoints does not cause problems
-software and hardware breakpoints work as expected
-watchpoints work as expected