Fix pyusb dependency
Installation of pyOCD fails at the time of this patch on platforms that
use pyusb, like Linux.  This is because the only released versions
of pyusb are alpha or beta, which pip does not install by default.
This patch sets the minimum version to the latest beta so pyusb
can be installed.