Improve callstack shown on crash
This patch adds try except blocks in the uninit call of the
board class around operations which access the target, transport
and interface.  This prevents exceptions from these calls from
replacing the current exception causing the uninit.  This patch
also prints the callstack of any of these calls if an error occurs.