Update pywinusb backend and add timeout
Update the pywinusb backend to support a read timeout.  If no timeout
is passed in then the default of 1s will be used.  This allows pyOCD to
gracefully fail, rather than lock up when a packet is dropped either
due to a pc problem or a firmware problem.

This patch also replaces the receive list with a deque instead, which
supports atomic O(1) pushes and pops.