Update to latest pywinusb to bring in fixes
Switch to the latest pywinusb to bring in the following fixes:
-Fix intermittent python.exe crash on some usb operations
-Prevent multiple scripts from accessing the same board, since
    this leads to one of the scripts raising an exception.

Note:
The python.exe crash is reproducable with pywinusb with a version
before 0.4.0 by running the script blank_test.py.  The exception
from multiple scripts running on the same board is easily
reproduceable by running two scripts at the same time.
One example of this is running basic_test.py in one console,
and while the script is still running, using the command
"pyocd-gdbserver -l" in another. The script basic_test.py will
immediately fail.