Refactor pyOCD in preperation for pyDapLink
Prepare pyOCD for daplink by using pyDapLink's interface.
Prior to this patch the board class depending on having both a
transport and an interface.  This caused the code to be more
tightly coupled to interfaces than it could be.

This patch combines most of the logic from the interface and transport
class and moves it into the DapLink class.  The DapLink class is an
implementation of the Link class, which provides all the functions
needed to communicate with a board.

More specific changes:
-Raw packet transfer test removed from speed test since pyDapLink
    will now be responsible for testing this
-References to 'transport' changed to 'link'
-Transport interface removed
-Interface code (usb drivers) moved into the transport folder to
    further isolate it from the rest of the codebase
-File cmsis_dap.py renamed to dap.py and moved into target folder
    for use by cortex_m.py and target_kinetis.py.  In the future
    dap.py will be removed and AP/DP drives will be used in
    its place.