Rename pyDAPLink to pyDAPAccess and modularize
Rename the transport folder to pyDAPAccess since this will become a new
module.  The name originally  planned for this new module was
pyDAPLink.  To reduce confusion with the DAPLink firmware project,
the new python module will be named  pyDAPAccess instead of pyDAPLink.

The name changes made in the patch are:
-Link -> DAPAccessIntf
-DapLink -> DAPAccessUSB
-transport folder -> pyDAPAccess folder

When pyOCD is ready to use the official pyDAP module the only changes
that should be necessary are as follows:
-Remove the pyDAPAccess folder
-Add pyDAPAccess as a dependency
-Rename imports from pyOCD.pyDAPAccess to just pyDAPAccess