Fix bug with missing product string on Linux
If a device without a product string is connected to a machine running
Linux then pyDAPAccess will not work correctly.  This is because the
product is set to 'None' which causes a crash when find is called on
it.  This patch fixes that problem by checking to make sure the product
string is not 'None' before using it.