Fix hanging GDB when continuing after a hardfault
If a hard fault occurs while debugging the signal SIGSEGV is returned
on halt.  When trying to continue after this hard fault the GDB client
sends the vCommand "Cont;C0b:1;c".  This is ignored in the vCont
processing since nothing matches the action "C0b".  This causes the
target to remain halted and the gdb client and server to go out of
sync.

This patch fixes this problem by comparing against only the first
character in the command.  That way if a continue or step packet
comes in with a signal specified then the correct action will still be
taken.  This patch also adds an error log statement to print if there
is an unsupported command.