Support accessing multiple devices simultaneously
Update the code to ensure listing the available boards does not
interfere with other pyOCD threads or processes listing or using
boards at the same time. Ensure opening and using a board does
not interfere with other threads or processes that are listing boards
or using boards (aside from the opened one).

Changes this patch makes:
-chooseBoard only ever locks the board it selects rather than all boards
-The DAPAccessCMSISDAP open function only locks the board it uses
-The function getAllConnectedInterface now returns closed handles
    to devices rather than the previous open handles.

Note:
Due to deficiencies in hidapi on Windows boards cannot be locked
for exclusive access.