Add pack-target support to tools
Allow cmsis-pack targets to be be used with all pyOCD tools -
pyocd-gdbserver, pyocd-flashtool and pyocd-tool. Also allow
supported pack targets to be listed from pyocd-gdbserver and
pyocd-flashtool with the --list-targets command line option.

Note - Outputting all targets in json for use in eclipse makes
eclipse unusable. To prevent this only built-in targets are returned
when printing in json mode.