RTOS: Make vector catches show originating thread
This follows discussions with Alan Hayward, and research into the
issue of showing HardFaults etc decently.

There has been discussion and attempts to patch GDB to let it unwind
from the Main stack to the Process stack, so that when stopping inside a
HardFault, we could see where the fault had come from. Discussion and
links here:

    https://bugs.launchpad.net/gcc-arm-embedded/+bug/1566054

After thought and discussion with Alan, I've concluded that that
approach is not conceptually correct - it would better suit GDB's
target process model to present the Handler Mode as a distinct thread
from the Processes, and ensure they accurately show the full system
state.

PyOCD's RTOS awareness already effectively achieves this, if you have a
supported RTOS, and my previous patches have helped ensure that the view
of the current thread is as up-to-date as possible when inside an
exception.

In future it would make sense to have a "bare metal RTOS" provider
for pyOCD to present Process and Main stacks as 2 threads, so the
same effect is achieved without an actual RTOS.

This patch is an extra refinement step, which further special-cases the
ARM vector catch stop when you have separate Process and Handler
threads.

Before the patch, when stopping due to a HardFault vector catch, you
would see:

   Thread #5 stopped due to SIGSEGV at HardFault_Handler
   #2 main thread (Running), PC at exception site
   #3 timer thread (Wait[MsgQueue])
   #4 idle thread (Ready)
   #5 Hander Mode thread (HardFault), PC at HardFault_Handler [GDB current]

This patch now gives you a more GDB-conventional view:

   Thread #2 stopped due to SIGSEGV at fault location
   #2 main thread (Running), PC at exception site [GDB current]
   #3 timer thread (Wait[MsgQueue])
   #4 idle thread (Ready)
   #5 Hander Mode thread (HardFault), PC at HardFault_Handler

This view better follows documented GDB behaviour that the "signal" is
reported at the site it occurred. The conventional behaviour is that the
system stops with PC showing the thread that received the signal,
pointing at the instruction we would execute next in the foreground -
for a SIGSEGV this would be the exception-generating instruction itself,
which would be retried if the signal handler (or ARM exception handler)
returned.

The ARM system state isn't modified - at the moment of stop the Handler
Mode thread has come into existence, and its PC still shows the first
instruction of HardFault_Handler.

On subsequent steps, we will step into the handler, and the Handler Mode
thread will be current, just as on a Unix system we might step into the
signal handler (albeit in the same thread, rather than a separate
Handler thread).

This patch only refines behaviour when an exception arises from the
Process stack. If the exception actually happens Handler mode, or in
Thread mode using the Main stack you will still see "SIGSEGV at
HardFault_Handler". But at least you will be in the correct thread - the
site will be 2 frames up the backtrace.