RTOS: Correct stack offset calcs
Stack offset computation for exception register access was incorrect
when FPU was active.

Calculation was subtracting "SpExceptionOffset" (ie "amount pushed by
hardware") from the register offsets in the table, when it needed to be
subtracting the amount pushed by the OS, to account for the fact that
the OS hasn't done its software push.

For non-FPU cases, this error didn't affect the result, as both software
and hardware push 8 words, but it produced the wrong answer with FPU.

Attempt to clarify calculations by replacing `realSpOffset` and
`realSpExceptionOffset` with `swStacked` and `hwStacked`.