RTOS: Tidy SP read out and don't mess with PSP
Now that "unavailable register" logic is moved, neatest to combine the
special SP handling.

Stop messing with PSP to make it equal SP in the exception case.

SP does need to be adjusted, and be treated as a per-thread register to
enable stack unwinding, but PSP is conceptually a system register. Make
it pass through the thread awareness untouched - this may aid debugging
in the event the thread awareness goes awry.