RTX5/FreeRTOS: Go for live LR on vector catch
RTX5 awareness was always reading LR EXC_RETURN value from OS, but this
isn't necessarily up-to-date. As a special case for the common case of
stopping on a HardFault etc, detect vector catch and take live LR instead.

FreeRTOS expects LR to be below the hardware-stored registers, but again
this won't be there at the instant of vector catch, so take live LR.

Argon and Zephyr do not manipulate EXC_RETURN, so no changes needed.