Keep listening sockets open
Standard socket glue unnecessarily closed the listening socket whenever
closing the connection socket. Separate this out so that the abstract
socket only closes the listening socket on a new `cleanup()` call.

This means that we are sure we will retain whatever port was manually
or automatically selected on initialisation if persisting across
multiple GDB or telnet sessions.

This should have no effect on normal use (clients connect only when
idle), but will change behaviour if clients are queueing waiting for the
current session to end. Previously clients would have been queued
waiting for an accept (maybe 5, if the listen backlog parameter was
obeyed), but everyone in the queue would have all been rejected when the
active session was closed, and the listener socket got closed too. Only
the first client to attempt a new connection after the previous session
ended would get in.

(If intending "only able to connect while idle" semantics, the correct
thing to do would have been to close the listening socket as soon as a
connection was /accepted/).

Now, people connecting while a session is active will be queued for the
next session. But as we don't really want a large queue for the next
session, reduce the listen backlog to 1. Depending on system, this may
mean the second waiter will get immediate connection refusal.