Make thread contexts support doubles
Double float registers are now handled correctly by thread provider
contexts - the corresponding single float registers are now looked up in
the register offset tables.

To make this change feasible, the register offset table lookup is now
centralised in a helper which takes the lookup table(s) and a list of
special-case registers.