Converted trace log flags to trace loggers.
Multiple modules had LOG_X flags that were used to control trace log
output that normally doesn't need to be seen. These flags were replaced
with .trace child loggers of the module logger, with a default level
of CRITICAL so they won't output anything. The trace logger level can
be overridden with a logging config if trace output is required. The
custom log handlers used for DP/AP trace logging can also be replicated
with a logging config, so it was also removed.