many fixes for ITM/SWO/SWV
This work addresses a total of 9 issues that were preventing ITM + SWO
from working on pyOCD + LPC55S69 + CMSIS-DAP.  Some of these are
particular to our use case, but many are generic to pyOCD + SWO.

1. When ITM is enabled, TPR is incorrectly modified to *disable*
unprivileged writes when it intended the exact opposite, causing
unprivileged stimulus writes to be silently dropped.  This work
modifies TPR as intended:  to allow all unprivileged stimulus writes.

2. The remote probe interface doesn't correctly return SWO data;
the swo_read fails on a bad JSON encoding.  This work correctly returns
a list of integers representing bytes.

3. On CMSIS-DAP, when enabling SWO when SWO is already enabled, SWO is
not first disabled, causing subsequent configuration to fail and the
SWO reader to malfunction.  This work always disables SWO before
configuring it.

4. On CMSIS-DAP, the size sent in a DAP_SWO_DATA request does not
correctly adjust for the protocol overhead, potentially causing the
last four bytes of the SWO packet to be left unread.  These trailing
bytes will then be read on a subsequent read, causing an exception
when the data payload does not constitute the expected bytes.  This
work adjusts the size sent in DAP_SWO_DATA to reflect the protocol
overhead.

5. On CMSIS-DAP, both CommandError and DeviceError are constructed
without error messages, making it hard to debug failures from a stack
trace.  This work modifies the construction of these errors to always
specify an error message.

6. For reasons that are unclear, the SWO parser silently drops any ITM
instrumentation packet for any stimulus port other than port 0.  This
work modifies the SWO parser to not drop any ITM packets.

7. The GDB server creates a separate SWO reader for each core (even
though there is nothing core-specific with respect to SWO).  Because
there is no locking between these threads, having multiple threads
reading SWO causes more or less causes rampant corruption of the ITM
stream on multicore targets.  This work creates only a single SWO
reader thread, regardless of the number of cores.

8. There is no serialization between the SWO reader thread and the GDB
server.  As a result, if SWO and GDB are both simultaneously active,
both the SWO reader and the GDB server are likely to experience
protocol errors.  (This is likely the cause of issue #855.) This work
adds a very coarse grain locking scheme between the two to assure that
only one thread is talking to the underlying probe at at a time.

9. On the LPC55, resets with SWO enabled will cause ITM stimulus
writes in the target to hang:  because ITM is enabled but TRACECLKDIV
and TRACECLKSEL are reset, the stimulus port is not disabled but the
FIFO never becomes ready.  This work explicitly sets TRACECLKDIV and
TRACECLKSEL on a reset on the LPC55.