Fallback to pyusb as the default backend.
This will try to use `pyusb` as the default USB Interface backend.
So far only selected list of OS has this parameter set.
This prevents pyOCD working on a platforms that are not on the list.
`pyusb` may be used as "fallback" in this case as most versatile solution.
If `pyusb` is not available or fails to work we will know that explicitly.

Signed-off-by: Tomasz 'CeDeROM' CEDRO <tomek@cedro.info>