Gdbserver fault resilience (#1244)
* gdbserver: fixed some simple type errors.

* gdbserver: retry with timeout for faults waiting for target to stop running.

After getting a transfer error or fault inside GDBServer.resume()
while checking the target status, a timeout timer is started. A
successful read of target status disables the timeout. If the
timeout expires without a successful read, the resume() call exits
and tells gdb the target got an error.

Added a 'debug.status_fault_retry_timeout' option to control the
retry timeout in seconds. The default is 1 second, so the user-
visible behaviour is currently very similar to before this change.