Fix ROM table negative addresses (#1267)
* coresight: rom_table: adjust component addresses that go negative.

This is for cases where the address is intended to wrap around from
the ROM table's base. For instance, the Microchip SAML11's root ROM
table is at 0x41003000, containing an entry with offset -0x60f04000
to the standard Cortex-M23 ROM table at 0xe00ff000. Without the
adjustment, the M23 table base is -0x1ff01000.
Somewhat magically, it still worked without this adjustment, as the
Python bit-and operator converts to unsigned.

(Note: does not include added debug logs from #1267.)