CMSIS-DAP: pydapaccess: hidapi_backend: read thread, max packet count limit.
This change introduces a background read thread for the HidApiUSB
backend for CMSIS-DAP. Similar to the other backends, it reads and
queues incoming command responses as soon as they are available.
Also added a packet count maximum to deal with the arbitrary limit
of 30 queued IN reports in the Mac version of hidapi.