lists: StubProbe class for creating target type class instances.
SoCTarget now requires a non-None session.probe when it is constructed.
ListGenerator.list_targets() needs to instantiate (but otherwise not
use) the target type classes to extract data such as vendor, SVD file
location, etc, but doesn't have a probe to use. This is solved by defining
a StubProbe(DebugProbe) class that implements only the unique_id property
accessed by Session's ctor. This is just enough to allow target
instantiation.

Added a note to SoCTarget's doc comment about the necessity to not
try to use session.probe for much of anything in the ctor because of
this use case.