user scripts: print proxy.
The print() builtin is overridden in the user script namespace with
an instance of the new PrintProxy class, primarily to allow redirecting
of print() calls to a CommandExecutionContext's output stream. This
ensure that prints from functions and commands in user scripts will
be sent over gdbserver monitor command output. The print proxy
is available via the user_script_print_proxy property on Session.