coresight: don't read CoreSight registers for non-CoreSight components (#1213)
The CoreSight spec (IHI0029D B2.3) says "All component classes require the implementation
of the Component and Peripheral Identification registers", but non-CoreSight components
may sometimes fault on reading registers outside this space. In particular, a Cortex-M7
SCS, while the core is in reset, will fault on an attempt to read 0xe000efbc, which relates
to the data cache and not component identification.

To fix, we should avoid reading registers outside PIDR and CIDR unless the component
has the CoreSight component class.