rtos: Fix reading Zephyr thread name
The Zephyr thread structure was changed to hold a buffered copy of the
thread name rather than a pointer to the thread name, in order to fix
potential access issues when reading from user mode. This was
implemented in Zephyr v2.0.0 and backported to the 1.14 LTS branch.

See https://github.com/zephyrproject-rtos/zephyr/pull/17071 for further
details.

Signed-off-by: Maureen Helm <maureen.helm@intel.com>